Program Skeleton
Here is the start of what I believe the program should look like. I have made lots of notes to make it as easy as possible for everyone to see what's going on and what needs to be done next. Inside the main loop of the Simulator class I have described the basic code that needs to go in there. The state of all the queues also needs to be stored in some way every iteration of the loop so I can record it using the DataCollector class.